Move exceptions to a single module
