+++ push_notifications/admin.py
@@ -3,11 +3,9 @@ from django.contrib import admin, messages
 from django.utils.encoding import force_str
 from django.utils.translation import gettext_lazy as _
 
-from .apns import APNSServerError
-from .gcm import GCMError
+from .exceptions import APNSServerError, GCMError, WebPushError
 from .models import APNSDevice, GCMDevice, WebPushDevice, WNSDevice
 from .settings import PUSH_NOTIFICATIONS_SETTINGS as SETTINGS
-from .webpush import WebPushError
 
 
 User = apps.get_model(*SETTINGS["USER_MODEL"].split("."))
