+++ push_notifications/gcm.py
@@ -11,7 +11,7 @@ from django.core.exceptions import ImproperlyConfigured
 
 from .compat import Request, urlopen
 from .conf import get_manager
-from .exceptions import NotificationError
+from .exceptions import GCMError
 from .models import GCMDevice
 
 
@@ -28,11 +28,6 @@ FCM_NOTIFICATIONS_PAYLOAD_KEYS = [
 	"title", "body", "icon", "sound", "badge", "color", "tag", "click_action",
 	"body_loc_key", "body_loc_args", "title_loc_key", "title_loc_args", "android_channel_id"
 ]
-
-
-class GCMError(NotificationError):
-	pass
-
 
 def _chunks(l, n):
 	"""
