--- tests/_mock.py
@@ -1,4 +1,0 @@
-try:
-	from unittest import mock  # noqa
-except ImportError:
-	from mock import mock  # noqa
