+++ tests/test_apns_push_payload.py
@@ -1,9 +1,10 @@
+from unittest import mock
+
 from apns2.client import NotificationPriority
 from django.test import TestCase
 
-from push_notifications.apns import APNSUnsupportedPriority, _apns_send
-
-from ._mock import mock
+from push_notifications.apns import _apns_send
+from push_notifications.exceptions import APNSUnsupportedPriority
 
 
 class APNSPushPayloadTest(TestCase):
