+++ tests/test_gcm_push_payload.py
@@ -1,8 +1,9 @@
+from unittest import mock
+
 from django.test import TestCase
 
 from push_notifications.gcm import send_bulk_message, send_message
 
-from ._mock import mock
 from .responses import GCM_JSON, GCM_JSON_MULTIPLE
